# o3api/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger REST API for O3as
consumes:
  - "application/json"
basePath: "/api"
...
paths:
...
  /get_model_info:
    post:
      operationId: "o3api.api.get_model_info"
      tags:
        - "get_model_info"
      summary: "Returning information about a model"
      description: "Return model info"
      produces:
        - "application/json"
      parameters:
        - name: ptype
          in: query
          type: string
          description: Plot type (tco3_zm, vmro3_zm, ...)
          default: "tco3_zm"
          required: true
        - name: model
          in: query
          type: string
          description: model name
          default: CCMI-1_ACCESS-refC2
          required: true
      responses:
        200:
          description: "Successfully returned model information"
          schema:
            $ref: '#/definitions/ModelInfo'
...
definitions:
...
  ModelInfo:
    type: object
    properties:
      name:
        type: string
      attrs:
        type: string
      dims:
        type: string
      coords:
        type: string